backend_python/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
supabase
